sinksmtp.go: pull out the dot-suffix-generation bit of address lists
This is the code that generates '.b.c' and '.c' from 'a.b.c'. It's been
pulled out because I'm going to need to reuse it soon for matching EHLOs.
In the process it became simplest to support '@' as 'match everything',
although I may change my mind about that later.