sinksmtp.go: we now verify reverse DNS results and put them in message metadata
We still log the unverified reverse DNS as before, but now we also log
the DNS results that are verified. In the future we may split logged DNS
results up further.

As part of this, we now sort reverse DNS results into a consistent order.
Before our reverse DNS came out in whatever order the resolver gave
it to us, with the result that two connections from the same remote IP
could have a different order of reverse DNS (which would have affected
message hashes).