sinksmtp: fully transition to fully working rules
All of sinksmtp's various rejection and processing options now actually
create rules under the hood. This has enabled the deletion of a whole
pile of code and makes the core logic a lot simpler. A number of bugs
were shaken out in the process of making this switch, which is good.

This also brings us up to what should be a fully working and more or
less finalized rule engine and rule language.