sinksmtp: correctly handle rules that match RCPT TO but defer to later
'@data reject to a@b' must be evaluated at RCPT TO time for each
RCPT TO value (unless I gorily change the semantics) but only take
effect at DATA (the same issue holds for @message). We now handle
this corner case correctly, or at least I hope.

This was a pain in the rear.

(I actually could change the semantics just in Decide(), by looking
for rules that require pRto and then running them once for each of
the accepted recipients. The rule would match if it matched for any
recipient. Hmm. That actually has better semantics than the current
setup; it is a bit weird that a @data time rule can match against a
rejected RCPT TO address.)