sinksmtp: improve rules parsing error message in a confusing case
'reject host something or bogus bogus' used to complain that the
right side of the OR was empty. This is technically correct but
very misleading to read; the right side of the OR is empty not
because we ran out of things to parse but because 'bogus' is not
a match operator. Now we generate the right error so I will be
less confused.