sinksmtp: move all user documentation into cmd/doc.go
The 'manual page' from sinksmtp.go moves here, as does the documenation
in RULES, and both got reformatted to look nice in godoc. This allows
godoc to show everything because having godoc show everything is the Go
style. cmd/rules.go was also revised to remove its out of date copy of
the control rules.