smtpd.go: reform how we get various connection parameters
Because it was getting unwieldy we now have a Config structure that
contains all of the various settings (plus some new ones) and this
struct is passed to NewConn. This eliminates a number of method
functions that were effectively setters for stuff that is now covered in
Config.

sinksmtp has been revised to match the new state of affairs.