sinksmtp: support multiple TLS certificates at once
This adds support for specifying multiple TLS certificates so that
crypto/tls can pick the right one if the client uses SNI. This is
supported on both the command line and in the -conncfg config file.

As a minor side change, sinksmtp also saves the server name that the
client sent us if it did and has been updated for smtpd.go's change in
how TLS connection information is exposed.