sinksmtp: fix a bug with matching DNS names due to trailing '.'
Go's resolver sensibly adds a trailing '.' on the end of resolved
DNS names for, eg, PTR lookups. Unfortunately our host matching code
wasn't dealing with this, with the result that basically all 'host ...'
statements would fail to match.

(They passed tests because I had a synthetic DNS lookup result that
also lacked the trailing '.'. This shows the perils of tests with
synthetic data.)