sinksmtp.go: support selectively disabling TLS in rules + a bit better TLS
This adds a new @connect phase and a 'tls-opt' setting for with.
The latter allows both disabling TLS and disabling client certificate
verification.

Sinksmtp also now disables asking for client certificates if the
source IP has already failed TLS setup once and this is the last
chance. This gives us a chance to set up TLS with clients that
have bad client certs but no other problems.