sinksmtp.go: count RSETs against do-nothing client settings if applicable
Until now, sinksmtp's detection of do-nothing clients only kicked in
if a client actively disconnected and then reconnected. If a client fed
us thousands of unsuccessful delivery attempts in a single connection,
RSET'ing between each, we did nothing. This changes that behavior so
that now RSETs count the same as disconnects.

In case you're curious: yes, there are clients that do this. Very
rare ones, but I ran into one.

(Obviously this only counts if you have set -minphase to 'from' or
later.)