sinksmtp: improve handling of 'EHLO .', including not panicing sometimes
The first improvement is that having a rule of eg 'helo .something'
will no longer panic with a slice indexing error in sDotIter.Next()
in this case.

We now support a new helo-has attribute of 'bogus', which is true for
such HELO/EHLOs. They are also considered to be 'nodots', for reasons
involving hand-waving.

Additional tests have been added for various aspects of this.

In the process, 'with note "..."' now correctly forbids notes with
embedded newlines. This was previously missed because my test case
for this was bad.