Add various expvar-exported statistics
These stats are what I'm interested in given my usage of sinksmtp as a
spam sink and spam rejector, but they may not be quite what other
people will find useful. The exact meaning of the statistics is not
currently documented apart from through the code and the
namespaces (and names) may not be entirely ideal and is subject to
change at a later date.

Some of the implementation is a bit of a hack, since we reach into
various spots in order to pull out information (or to put it into the
stats).