Explicitly check for the RFC 7505 null MX
We were already rejecting 'MX .' as a bogus MX entry, but now we
specifically check for 'MX 0 .' and note it as a RFC 7505 null MX.

I've added some tests for null and bogus 'localhost.' MX entries, but
they rely on specific outside domains having these MX entries, which
is out of my control. Doing better is unfortunately a lot more
complicated, although theoretically possible by being excessively
clever with an interface for a net.Resolver.