rlex*.go: don't shadow a type with a variable name
I was using 'item' as both a type and a variable name, as found by 'go
vet -shadow'. This is probably a bad idea, especially in the context
of a function that returns something of type 'item' and was using
'item' as a variable name for what it returned. Now I've renamed the
'item' variables to 'itm'.

(Perhaps this shows that the type should be renamed, but that would be
a much bigger project and I'm bad with names.)