sinksmtp: fsync the just-written message file and check for errors
We now call .Sync() on the file after we've written it, and check for
errors on both .Write() and .Close(). In theory calling .Sync() on the
file alone is not sufficient; to be fully compliant we would open the
directory and also .Sync() it. But no, that's too much work in order
to deal with Linux filesystem people.