Correctly format accept, reject, and tempfail messages
Under some circumstances, the raw messages we use may contain '%'
characters in situations where they would be valid fmt control codes,
which would cause the messages to be mangled when smtpd treats the
messages as format strings. Switch to explicitly formatting them with
"%s", even though this is a slightly silly way of formatting a single
string.