Ask clients for a TLS certificate and record information about it
Increasingly, SMTP senders are willing to offer a client TLS
certificate to you if you ask. We ask, then check to see if any
provided TLS certificate verifies, and record the results in both the
log and the received message data (if either is applicable).

We directly perform verification rather than setting .ClientAuth to
tls.VerifyClientCertIfGiven, because we still want to accept the
connection if the client's TLS certificate fails to verify.